test: step1과 step2 테스트 분리
